// src/types/Car.ts
export type Car = {
	carId: number;
	catalogId: number;
	makeId: number;
	model: string;
	color: string;
	fuel: string;
	transmission: string;
	driveType: string;
	capacity: number;
	engineSize: number;
	numberOfDoors: number;
	vinNumber: string;
	price: number;
	status: string;
	year: number;
	description: string;
	image: string;
};